+++ lib/src/home/view/screens/dart_packages.dart
@@ -31,7 +31,7 @@ class DartPackages extends StatelessWidget {
             shrinkWrap: true,
             itemCount: con.packages.length,
             gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-              crossAxisCount: MyApp.inSmallScreen ? 2 : 3,
+              crossAxisCount: App.inSmallScreen ? 2 : 3,
               childAspectRatio: 5 / 4,
               crossAxisSpacing: 10,
             ),